Move wrapping OpenTelemetry to LibraryTestRunner
